fix: ignore letter casing for column types
As part of the underlying request from the plugin to the SQLite database
the types of columns need to be converted. The inference based on the
returned column types now ignores the case of the type.
This needs to be done as the column type is returned as it was entered
during the table creation